requirements.txt
requests==2.25.1
aliyun_python_sdk_alidns==2.6.19
aliyun_python_sdk_core==2.13.30
PyYAML==5.3.1
